Move dimension_name lookup to TreeService
