+++ transmart-core-db/grails-app/services/org/transmartproject/db/tree/TreeService.groovy
@@ -14,6 +14,8 @@ import org.transmartproject.core.ontology.OntologyTermTag
 import org.transmartproject.core.ontology.OntologyTermTagsResource
 import org.transmartproject.db.accesscontrol.AccessControlChecks
 import org.transmartproject.db.i2b2data.Study
+import org.transmartproject.db.metadata.DimensionDescription
+import org.transmartproject.db.multidimquery.DimensionImpl
 import org.transmartproject.db.multidimquery.QueryService
 import org.transmartproject.db.multidimquery.query.Constraint
 import org.transmartproject.db.multidimquery.query.ConstraintFactory
@@ -67,6 +69,7 @@ class TreeService {
                         currentNode,
                         children
                 )
+                node.dimension = getDimension(node.tableName, currentNode.code)
                 node.children.each { child ->
                     child.parent = node
                 }
@@ -88,6 +91,28 @@ class TreeService {
 
     static Criterion contains(String propertyName, String value) {
         StringUtils.like(propertyName, value, MatchMode.ANYWHERE)
+    }
+
+    static String getDimension(String table, String modifierCode) {
+        switch (table) {
+            case 'concept_dimension':
+                return DimensionImpl.CONCEPT.name
+            case 'patient_dimension':
+                return DimensionImpl.PATIENT.name
+            case 'modifier_dimension':
+                def dimension = DimensionDescription.createCriteria().list {
+                    eq('modifierCode', modifierCode)
+                } as List<DimensionDescription>
+                if (dimension.size() > 0) {
+                    return dimension.first().name
+                }
+                break
+            case 'trial_visit_dimension':
+                return DimensionImpl.TRIAL_VISIT.name
+            case 'study':
+                return DimensionImpl.STUDY.name
+        }
+        return 'UNKNOWN'
     }
 
     /**
@@ -200,5 +225,4 @@ class TreeService {
         }
         forest
     }
-
 }
