+++ transmart-core-db/src/main/groovy/org/transmartproject/db/tree/TreeNode.groovy
@@ -4,8 +4,6 @@ import groovy.transform.CompileStatic
 import org.transmartproject.core.ontology.OntologyTerm
 import org.transmartproject.core.ontology.OntologyTermTag
 import org.transmartproject.core.ontology.OntologyTermType
-import org.transmartproject.db.metadata.DimensionDescription
-import org.transmartproject.db.multidimquery.DimensionImpl
 import org.transmartproject.db.multidimquery.TrialVisitDimension
 import org.transmartproject.db.multidimquery.query.Combination
 import org.transmartproject.db.multidimquery.query.ConceptConstraint
@@ -65,10 +63,6 @@ class TreeNode {
 
     final String getColumnName() {
         delegate.columnName.toLowerCase().trim()
-    }
-
-    final String getModifierCode() {
-        delegate.code
     }
 
     /**
@@ -274,28 +268,5 @@ class TreeNode {
         } else {
             OntologyTermType.UNKNOWN
         }
-    }
-
-    final String getDimension() {
-        switch (tableName) {
-            case 'concept_dimension':
-                return DimensionImpl.CONCEPT.name
-            case 'patient_dimension':
-                return DimensionImpl.PATIENT.name
-            case 'modifier_dimension':
-                def dimension = DimensionDescription.createCriteria().list {
-                    eq('modifierCode', modifierCode)
-                } as List<DimensionDescription>
-                if (dimension.size() > 0) {
-                    return dimension.first().name
-                }
-                break
-            case 'trial_visit_dimension':
-                return DimensionImpl.TRIAL_VISIT.name
-            case 'study':
-                return DimensionImpl.STUDY.name
-
-        }
-        return 'UNKNOWN'
     }
 }
