+++ transmart-rest-api/src/main/groovy/org/transmartproject/rest/marshallers/TreeNodeMarshaller.groovy
@@ -37,10 +37,12 @@ class TreeNodeMarshaller implements ObjectMarshaller<JSON> {
                 name:     obj.name,
                 fullName: obj.fullName,
                 type:     obj.ontologyTermType.name(),
-                dimension: obj.dimension,
                 visualAttributes: obj.visualAttributes
         ] as Map<String, Object>
         def constraint = obj.constraint
+        if (obj.dimension) {
+            result.dimension = obj.dimension
+        }
         if (constraint) {
             result.constraint = constraint
         }
